Set svn:ignore to include *.d Make all the silly warnings go away. ;-) Also, possibly fixed a bug relating to an uninitialized variable in handle_welcome().

+++ src/dns.c
@@ -56,7 +56,7 @@ char *dns_srv_list(char *hostname) {
 
 
 	n_srv = 0;
-	while(h->ancount-- && n_srv < sizeof(svr)/sizeof(svr[0])) {
+	while(h->ancount-- && n_srv < (int)(sizeof(svr)/sizeof(svr[0]))) {
 		if((hlen = dn_expand(answer, answer+alen, p, host, sizeof(host))) < 0)
 			break;
